Move cleanup to one place (#171)
* Removed more Apache Http clients and move idam information to http context

* Format fixes

* Begin moving answer cleanup

* Start moving questions cleanup

* Move cleanup to one place

* Better class hierarchy

* Not allowed to inherit from Step

* Remove duplicate delete

* Remove duplicate code

+++ src/aat/java/uk/gov/hmcts/reform/coh/functional/bdd/steps/ApiSteps.java
@@ -31,8 +31,7 @@ import uk.gov.hmcts.reform.coh.functional.bdd.utils.TestContext;
 import uk.gov.hmcts.reform.coh.functional.bdd.utils.TestTrustManager;
 import uk.gov.hmcts.reform.coh.repository.*;
 import uk.gov.hmcts.reform.coh.schedule.notifiers.EventNotifierJob;
-import uk.gov.hmcts.reform.coh.service.OnlineHearingService;
-import uk.gov.hmcts.reform.coh.service.SessionEventService;
+import uk.gov.hmcts.reform.coh.service.*;
 import uk.gov.hmcts.reform.coh.states.SessionEventForwardingStates;
 import uk.gov.hmcts.reform.coh.utils.JsonUtils;
 
@@ -74,6 +73,21 @@ public class ApiSteps extends BaseSteps {
     @Autowired
     private OnlineHearingRepository onlineHearingRepository;
 
+    @Autowired
+    private QuestionService questionService;
+
+    @Autowired
+    private AnswerService answerService;
+
+    @Autowired
+    private AnswerRepository answerRepository;
+
+    @Autowired
+    private DecisionService decisionService;
+
+    @Autowired
+    private DecisionReplyRepository decisionReplyRepository;
+
     private JSONObject json;
 
     private CloseableHttpClient httpClient;
@@ -105,29 +119,69 @@ public class ApiSteps extends BaseSteps {
 
     @After
     public void cleanUp() {
-        for (String caseId : caseIds) {
+        for (DecisionReply decisionReply : testContext.getScenarioContext().getDecisionReplies()) {
             try {
-                OnlineHearing onlineHearing = new OnlineHearing();
-                onlineHearing.setCaseId(caseId);
-                onlineHearing = onlineHearingService.retrieveOnlineHearingByCaseId(onlineHearing);
-                onlineHearingPanelMemberRepository.deleteByOnlineHearing(onlineHearing);
-                onlineHearingService.deleteByCaseId(caseId);
-            } catch (DataIntegrityViolationException e) {
-                log.error(
-                    "Failure may be due to foreign key. This is okay because the online hearing will be deleted elsewhere.");
+                decisionReplyRepository.deleteById(decisionReply.getId());
+            } catch (Exception e) {
+                log.error("Failure may be due to foreign key. This is okay because the online hearing will be deleted elsewhere.");
             }
         }
+
         if (testContext.getScenarioContext().getSessionEventForwardingRegisters() != null) {
-            for (SessionEventForwardingRegister sessionEventForwardingRegister : testContext.getScenarioContext()
-                .getSessionEventForwardingRegisters()) {
+            for (SessionEventForwardingRegister sessionEventForwardingRegister : testContext.getScenarioContext().getSessionEventForwardingRegisters()) {
                 try {
                     sessionEventForwardingRegisterRepository.delete(sessionEventForwardingRegister);
                 } catch (DataIntegrityViolationException e) {
-                    log.error(
-                        "Failure may be due to foreign key. This is okay because the online hearing will be deleted elsewhere.");
+                    log.error("Failure may be due to foreign key. This is okay because the online hearing will be deleted elsewhere.");
                 }
             }
         }
+
+        // Delete all decisions
+        if (testContext.getScenarioContext().getCurrentDecision() != null) {
+            Decision decision = testContext.getScenarioContext().getCurrentDecision();
+            try {
+                decisionService.deleteDecisionById(decision.getDecisionId());
+            }
+            catch (Exception e) {
+                log.debug("Unable to delete decision: " + decision.getDecisionId());
+            }
+        }
+
+        if (testContext.getScenarioContext().getCaseIds() != null) {
+
+            for (String caseId : testContext.getScenarioContext().getCaseIds()) {
+                try {
+                    OnlineHearing onlineHearing = new OnlineHearing();
+                    onlineHearing.setCaseId(caseId);
+                    onlineHearing = onlineHearingService.retrieveOnlineHearingByCaseId(onlineHearing);
+
+                    // Delete all the Q & A
+                    Optional<List<Question>> questionList = questionService.findAllQuestionsByOnlineHearing(onlineHearing);
+                    if (questionList.isPresent()) {
+                        for (Question question : questionList.get()) {
+                            List<Answer> answers = answerService.retrieveAnswersByQuestion(question);
+                            if (!answers.isEmpty()) {
+                                for (Answer answer : answers) {
+                                    answerRepository.delete(answer);
+                                }
+                            }
+                            questionService.deleteQuestion(question);
+                        }
+                    }
+
+                    // First delete event linked to an online hearing
+                    sessionEventService.deleteByOnlineHearing(onlineHearing);
+
+                    // Now delete the panel members
+                    onlineHearingPanelMemberRepository.deleteByOnlineHearing(onlineHearing);
+                    onlineHearingService.deleteByCaseId(caseId);
+                } catch (DataIntegrityViolationException e) {
+                    log.error("Failure may be due to foreign key. This is okay because the online hearing will be deleted elsewhere.");
+                }
+            }
+        }
+
         for (Jurisdiction jurisdiction : testContext.getScenarioContext().getJurisdictions()) {
             try {
                 jurisdictionRepository.delete(jurisdiction);
@@ -141,10 +195,10 @@ public class ApiSteps extends BaseSteps {
     @When("^a get request is sent to ' \"([^\"]*)\"' for the saved online hearing$")
     public void a_get_request_is_sent_to(String endpoint) throws Throwable {
         OnlineHearing onlineHearing = testContext.getScenarioContext().getCurrentOnlineHearing();
-        String url = baseUrl + endpoint + "/" + onlineHearing.getOnlineHearingId().toString();
-        HttpEntity<String> request = new HttpEntity<>("", header);
         RestTemplate restTemplate = getRestTemplate();
-        ResponseEntity<String> response = restTemplate.exchange(url, HttpMethod.GET, request, String.class);
+        HttpEntity<String> request = new HttpEntity<>("", header);
+        ResponseEntity<String> response = restTemplate.exchange(baseUrl + endpoint + "/" + onlineHearing.getOnlineHearingId().toString(), HttpMethod.GET, request, String.class);
+
         testContext.getHttpContext().setResponseBodyAndStatesForResponse(response);
     }
 
@@ -160,6 +214,7 @@ public class ApiSteps extends BaseSteps {
     public void a_post_request_is_sent_to(String endpoint) throws Throwable {
         RestTemplate restTemplate = getRestTemplate();
         HttpEntity<String> request = new HttpEntity<>(json.toString(), header);
+
         ResponseEntity<String> response = restTemplate.exchange(baseUrl + endpoint, HttpMethod.POST, request, String.class);
         testContext.getHttpContext().setResponseBodyAndStatesForResponse(response);
     }
